Add update_package.sh and checks for init.sh to be run from ws
Now init.sh needs to be run from a workspace. This allows us to check
for whether the user already has baxter model packages locally
installed. update_package.sh does comprehensive checking on this and
gives the user instructions on how to install things or rectify the
installation if only a subset of the needed packages are found.

- init.sh must be called from a valid workspace. Checks for src dir,
  CMakeLists.txt and ensures CMakeLists.txt is a top-level catkin
  variant.
- Installs lula packages in src/lula subdirectory of the workspace.
- Checks the workspace for baxter_description and rethink_ee_description
  packages.
  - If a full installation is found, asks user whether those should be
    used.
  - If a partial installation is found, prompts user on how to fix that.
  - If no installation is found, asks user whether it should install
    baxter_common (which is where those packages come from) in the
    src/lula subdirectory.
- Updates all packages in src/lula when updating.